'Solve' major bug while calling api
Noticed a problem on API call. When user clicks on a map position, the location state is updated, calling a useEffect to call the weather api and update it's card. But when calling it a second time and forth when clicking the map, the api call executed inside the useEffect that watch location change does not complete until the user clicks anywhere on the screen. It seems like the call stays in standby. To circumvent the problem, changed  the other api calls to the button handler that updates the location state