Add the edit button and it's function
Now Map may receive an location as props, if so, updates the current location object with the location received. At List, changed the useEffect that receives the return from map, check if the 'new' location id exists on the current list, if true updates it, else add to the list.